pyproject.toml
[tool.poetry]
name = "test-backend"
version = "0.0.1"
description = "Test Backend"
authors = ["Dermofet"]


[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.103.2"
uvicorn = "^0.23.2"
SQLAlchemy = "^2.0.22"
asyncpg = "^0.28.0"
alembic = "^1.11.3"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
python-dotenv = "^1.0.0"
celery = "^5.3.1"
redis = "^5.0.1"
loguru = "^0.7.2"

[tool.poetry.dev-dependencies]
mypy = "^1.6.0"
black = {extras = ["d"], version = "^23.9.1"}
isort = "^5.12.0"
pytest = "^7.4.2"
pytest-cov = "^4.1.0"
pytest-alembic = "^0.10.7"
pytest-asyncio = "^0.21.1"
pytest-trio = "^0.8.0"


[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
line_length = 120

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"